requirements.txt
# Audio recording
sounddevice>=0.4.6
scipy>=1.11.0

# Video / image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# AI background removal
rembg>=2.0.50
onnxruntime>=1.16.0   # use onnxruntime-silicon on Apple Silicon for CoreML acceleration

# Progress
tqdm>=4.65.0
